requirements.txt
kafka-python==2.0.2
pytest==6.2.5